# platform/backend/src/packages/application/api/api-swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/company/{companyId}/role/{userId}": {
            "get": {
                "operationId": "CompanyUserRoleGetController_executeExtended",
                "summary": "Get company user roles list",
                "parameters": [
                    {
                        "name": "companyId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCompany"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization failed"
                    },
                    "403": {
                        "description": "Forbidden resource"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "operationId": "CompanyUserRoleSetController_executeExtended",
                "summary": "Get company user roles list",
                "parameters": [
                    {
                        "name": "companyId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCompany"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization failed"
                    },
                    "403": {
                        "description": "Forbidden resource"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Karma Platform Backend API",
        "description": "Karma Platform API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "UserCompany": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}